% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena.R
\docType{class}
\name{AthenaDriver-class}
\alias{AthenaDriver-class}
\alias{AthenaDriver}
\title{Athena driver class.}
\usage{
\S4method{initialize}{AthenaDriver}(.Object, ...)
}
\description{
Athena driver class.

Constructor of AthenaDriver
}
\keyword{internal}
