% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena.R
\docType{methods}
\name{dbGetQuery,AthenaConnection,character-method}
\alias{dbGetQuery,AthenaConnection,character-method}
\title{Send query and retrieve results}
\usage{
\S4method{dbGetQuery}{AthenaConnection,character}(conn, statement, ...)
}
\description{
Send query and retrieve results
}
