% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena.R
\docType{methods}
\name{dbSendQuery,AthenaConnection,character-method}
\alias{dbSendQuery,AthenaConnection,character-method}
\title{Execute an Athena Query}
\usage{
\S4method{dbSendQuery}{AthenaConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{An Athena Connection}

\item{statement}{A SQL statement}

\item{...}{delegated to JDBC}
}
\description{
Execute an Athena Query
}
