% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena.R
\docType{methods}
\name{fetch,AthenaResult,numeric-method}
\alias{fetch,AthenaResult,numeric-method}
\title{Fetch Athena Results}
\usage{
\S4method{fetch}{AthenaResult,numeric}(res, n = -1, ...)
}
\arguments{
\item{res}{an AthenaResult}

\item{n}{-1 for all, or how many records to fetch}

\item{...}{delegated to JDBC}
}
\description{
Fetch Athena Results
}
